Renamed package `player` to `psync` ([p]layer [sync]hronization)
